rigol-ds: Add support for DSO3000 protocol
This is the protocol used by Agilent DSO3000 series (likely also Rigol
DS5000). Apart from tiny differences and different sample format, it
is mostly the same as VS5000.

The sample data sent by the scope seems to always be somewhat
garbled and Agilent Scope Connect (that seems to be a version of Rigol
Ultrascope for DS5000) silently tolerates it. This does the same thing.

Details about the protocol and the bugs is in a comment.