scpi_tcp: Fix completion of small reads
Calculation of response length in scpi_tcp_raw_read_data() doesn't look
right -- with a value of maxlen=1 scpi_tcp_read_complete() would never
return success.

I'm actually not sure why the maxlen is bumped by 1 in
scpi_tcp_raw_read_data(). My suspicion is that it's to prevent
signalling completion in case the response is unexpectedly long.
Let's not do that when the response is just the right size.